+++ include/Renderer.hpp
@@ -55,7 +55,12 @@ void draw(
     const Vec2& pivot = Anchor::CENTER
 );
 
-void draw(const Texture& texture, const Rect& dst);
+void draw(const Texture& texture, Rect dst);
+
+void draw9Slice(
+    const Texture& texture, Rect dst, const Rect& slice, const Vec2& anchor = Anchor::TOP_LEFT,
+    const Vec2& pivot = Anchor::CENTER
+);
 
 void drawBatch(
     const Texture& texture, const std::vector<Transform>& transforms,