+++ src/font.cpp
@@ -69,35 +69,35 @@ Font::~Font()
     }
 }
 
-void Font::setAlignment(const Align alignment) const
+void Font::setAlignment(const TextAlign alignment) const
 {
     switch (alignment)
     {
-    case Align::Left:
+    case TextAlign::Left:
         TTF_SetFontWrapAlignment(m_font, TTF_HORIZONTAL_ALIGN_LEFT);
         break;
-    case Align::Center:
+    case TextAlign::Center:
         TTF_SetFontWrapAlignment(m_font, TTF_HORIZONTAL_ALIGN_CENTER);
         break;
-    case Align::Right:
+    case TextAlign::Right:
         TTF_SetFontWrapAlignment(m_font, TTF_HORIZONTAL_ALIGN_RIGHT);
         break;
     }
 }
 
-Align Font::getAlignment() const
+TextAlign Font::getAlignment() const
 {
     const TTF_HorizontalAlignment align = TTF_GetFontWrapAlignment(m_font);
     switch (align)
     {
     case TTF_HORIZONTAL_ALIGN_LEFT:
-        return Align::Left;
+        return TextAlign::Left;
     case TTF_HORIZONTAL_ALIGN_CENTER:
-        return Align::Center;
+        return TextAlign::Center;
     case TTF_HORIZONTAL_ALIGN_RIGHT:
-        return Align::Right;
+        return TextAlign::Right;
     default:
-        return Align::Left;
+        return TextAlign::Left;
     }
 }
 
@@ -341,14 +341,9 @@ Raises:
 
         .def_prop_rw("alignment", &Font::getAlignment, &Font::setAlignment, R"doc(
 Get or set the text alignment for wrapped text.
-
-Valid values: Align.LEFT, Align.CENTER, Align.RIGHT
         )doc")
         .def_prop_rw("hinting", &Font::getHinting, &Font::setHinting, R"doc(
 Get or set the font hinting mode.
-
-Valid values: FontHinting.NORMAL, FontHinting.MONO, FontHinting.LIGHT,
-              FontHinting.LIGHT_SUBPIXEL, FontHinting.NONE
         )doc")
         .def_prop_rw("pt_size", &Font::getPtSize, &Font::setPtSize, R"doc(
 Get or set the point size of the font. Values below 8 are clamped to 8.